+++ src/main/java/org/truffleruby/extra/TruffleRubyNodes.java
@@ -29,7 +29,7 @@ import org.truffleruby.shared.BuildInformationImpl;
 @CoreClass("TruffleRuby")
 public abstract class TruffleRubyNodes {
 
-    @CoreMethod(names = "graal?", onSingleton = true)
+    @CoreMethod(names = "jit?", onSingleton = true)
     public abstract static class GraalNode extends CoreMethodArrayArgumentsNode {
 
         @TruffleBoundary
@@ -50,7 +50,7 @@ public abstract class TruffleRubyNodes {
 
     }
 
-    @CoreMethod(names = "sulong?", onSingleton = true)
+    @CoreMethod(names = "cexts?", onSingleton = true)
     public abstract static class SulongNode extends CoreMethodArrayArgumentsNode {
 
         @Specialization