+++ src/main/ruby/core/truffle/backward.rb
@@ -30,3 +30,17 @@ module Truffle
   end
 
 end
+
+module TruffleRuby
+
+  def self.sulong?
+    warn 'TruffleRuby.sulong? has been replaced by TruffleRuby.cexts? and will be removed at a future release', uplevel: 1
+    cexts?
+  end
+
+  def self.graal?
+    warn 'TruffleRuby.graal? has been replaced by TruffleRuby.jit? and will be removed at a future release', uplevel: 1
+    jit?
+  end
+
+end