+++ test/truffle/compiler/pe/pe.rb
@@ -26,7 +26,7 @@
 #
 #   jt ruby --graal --vm.Dgraal.TraceTruffleCompilation=true --vm.Dgraal.TruffleCompilationExceptionsAreFatal=true --vm.Dgraal.TruffleIterativePartialEscape=true test.rb
 
-unless TruffleRuby.graal?
+unless TruffleRuby.jit?
   puts 'not running the GraalVM Compiler'
   exit 1
 end